move and rename labelSelectionRenderer
