+++ src/components/commons/components/combo-box/selection/combo-box-label-selection.spec.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import LabelSelectionRenderer from './label-selection-renderer';
+import ComboBoxLabelSelection from './combo-box-label-selection';
 import { describe, it, expect } from 'vitest';
 describe('DefaultLabelRenderer', () => {
 	describe('DefaultLabelRenderer', () => {
@@ -12,7 +12,7 @@ describe('DefaultLabelRenderer', () => {
 
 		it('should render with selected option', () => {
 			const { getByText } = render(
-				<LabelSelectionRenderer
+				<ComboBoxLabelSelection
 					option={option}
 					placeholder="Select an option"
 				/>
@@ -24,7 +24,7 @@ describe('DefaultLabelRenderer', () => {
 
 		it('should render with placeholder', () => {
 			const { getByText } = render(
-				<LabelSelectionRenderer option={null} placeholder="Select an option" />
+				<ComboBoxLabelSelection option={null} placeholder="Select an option" />
 			);
 
 			expect(getByText('Select an option')).toBeInTheDocument();
@@ -32,7 +32,7 @@ describe('DefaultLabelRenderer', () => {
 
 		it('should render with search value', () => {
 			const { getByText } = render(
-				<LabelSelectionRenderer
+				<ComboBoxLabelSelection
 					option={null}
 					search="Search value"
 					placeholder="Select an option"
