+++ src/components/commons/index.ts
@@ -1,4 +1,3 @@
-export { default as LabelSelectionRenderer } from './components/label-selection-renderer';
 export { default as FieldContainer } from './components/field-container';
 export { default as Label } from './components/label';
 export { default as Fieldset } from './components/fieldset';
